requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
boto3==1.42.34
botocore==1.42.34
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.4
daphne==4.1.2
distro==1.9.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-extensions==4.1
django-storages==1.14.6
django-stubs-ext==5.2.8
django-tasks==0.11.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
gevent==25.9.1
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
icalendar==6.3.2
idna==3.11
Incremental==24.11.0
jmespath==1.1.0
kombu==5.6.2
msgpack==1.1.2
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.6
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
razorpay==2.0.0
redis==7.1.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.16.0
sentry-sdk==2.53.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.2
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==15.0.1
zope.event==6.1
zope.interface==8.2
